Better comments/instructions
When I first tried to use this role I was confused with what I was expected to configure. I assume that if I provide a netplan configuration file to `netplan_config_file` that the content of that file would be applied to the system. It took me a decent amount of time to figure out what the required options do. That is why I am sending this PR.